[1.3] Move Bucket class to Common
Move Bucket class from group_mod.py to common.py.
