+++ tests/v0x04/test_controller2switch/test_group_mod.py
@@ -1,7 +1,8 @@
 """group_mod tests."""
 from unittest import TestCase
 
-from pyof.v0x04.controller2switch.group_mod import Bucket, GroupMod
+from pyof.v0x04.controller2switch.common import Bucket
+from pyof.v0x04.controller2switch.group_mod import GroupMod
 
 
 class TestGroupMod(TestCase):
